+++ mayan/apps/logging/tests/test_models.py
@@ -1,15 +1,16 @@
 from mayan.apps.testing.tests.base import BaseTestCase
 
+from ..events import event_error_log_deleted
 from ..models import ErrorLogPartitionEntry
 
-from .mixins import ErrorLoggingTextMixin
+from .mixins import ErrorLogPartitionEntryTestMixin
 
 
-class ErrorLoggingModelTestCase(ErrorLoggingTextMixin, BaseTestCase):
+class ErrorLoggingModelTestCase(ErrorLogPartitionEntryTestMixin, BaseTestCase):
     def setUp(self):
         super().setUp()
-        self._create_error_log_test_object()
-        self._create_error_log_entry()
+        self._create_test_error_log_object()
+        self._create_test_error_log_entry()
 
     def test_entries_limit(self):
         self.error_log.limit = 3
@@ -23,6 +24,25 @@ class ErrorLoggingModelTestCase(ErrorLoggingTextMixin, BaseTestCase):
             ['2', '3', '4']
         )
 
-    def test_object_deletion(self):
+    def test_orphan_error_logs_after_object_deletion(self):
+        self._clear_events()
+
         self._test_object.delete()
         self.assertEqual(ErrorLogPartitionEntry.objects.count(), 0)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_error_log_deletion(self):
+        self._clear_events()
+
+        self._test_error_log_entry.delete()
+        self.assertEqual(ErrorLogPartitionEntry.objects.count(), 0)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_object)
+        self.assertEqual(events[0].target, self._test_object)
+        self.assertEqual(events[0].verb, event_error_log_deleted.id)