+++ mayan/apps/logging/views.py
@@ -1,17 +1,21 @@
 from django.contrib import messages
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.views.generics import ConfirmView, SingleObjectListView
+from mayan.apps.views.generics import (
+    ConfirmView, SingleObjectDeleteView, SingleObjectListView
+)
 from mayan.apps.views.mixins import ExternalContentTypeObjectViewMixin
 
 from .icons import icon_object_errors
 from .models import GlobalErrorLogPartitionEntry
-from .permissions import permission_error_log_view
+from .permissions import (
+    permission_error_log_entry_delete, permission_error_log_entry_view
+)
 
 
 class GlobalErrorLogEntryList(SingleObjectListView):
     model = GlobalErrorLogPartitionEntry
-    object_permission = permission_error_log_view
+    object_permission = permission_error_log_entry_view
 
     def get_extra_context(self):
         return {
@@ -31,7 +35,7 @@ class GlobalErrorLogEntryList(SingleObjectListView):
 class ObjectErrorLogEntryListClearView(
     ExternalContentTypeObjectViewMixin, ConfirmView
 ):
-    external_object_permission = permission_error_log_view
+    external_object_permission = permission_error_log_entry_delete
 
     def get_extra_context(self):
         return {
@@ -42,17 +46,40 @@ class ObjectErrorLogEntryListClearView(
         }
 
     def view_action(self):
-        self.external_object.error_log.all().delete()
+        self.external_object.error_log.clear(_user=self.request.user)
         messages.success(
             message=_('Object error log cleared successfully'),
             request=self.request
         )
 
 
+class ObjectErrorLogEntryDeleteView(
+    ExternalContentTypeObjectViewMixin, SingleObjectDeleteView
+):
+    external_object_permission = permission_error_log_entry_delete
+    object_permission = permission_error_log_entry_delete
+    pk_url_kwarg = 'error_log_partition_entry_id'
+
+    def get_extra_context(self):
+        return {
+            'external_object': self.external_object,
+            'navigation_object_list': ('external_object', 'object'),
+            'title': _('Delete error log entry: %s') % self.object
+        }
+
+    def get_instance_extra_data(self):
+        return {
+            '_event_actor': self.request.user
+        }
+
+    def get_source_queryset(self):
+        return self.external_object.error_log.all()
+
+
 class ObjectErrorLogEntryListView(
     ExternalContentTypeObjectViewMixin, SingleObjectListView
 ):
-    external_object_permission = permission_error_log_view
+    external_object_permission = permission_error_log_entry_view
 
     def get_extra_context(self):
         return {